end-to-end_project_3/pyproject.toml
[project]
name = "legalmind-ai"
version = "0.1.0"
description = "Multi-agent legal document analyst with RAG and LangGraph orchestration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # LLM & Orchestration
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "langgraph>=0.2.0",
    
    # Vector Store & Embeddings
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    
    # Document Processing
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    
    # Utilities
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",  # Pretty CLI output
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "ruff>=0.5.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
